[ChannelAdmin] fix GoogleAuth.tsx onDidMount logic
Previously, if the user was already signed in when the component
mounted, it would navigate them to the root dashboard screen.
E.g. if I was on /dashbaord/global-configs/SOME-CONFIG-ID, then
refreshed the page, <GoogleAuth /> would navigate me back to
/dashboard. This commit modifies the logic slighty so that no
navigation happens if the user is already logged on when the
component mounts.